chore : renamed TestProject -> dontworry
