+++ dontworry/config/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for TestProject project.
+Django settings for config project.
 
 Generated by 'django-admin startproject' using Django 4.2.3.
 
@@ -69,7 +69,7 @@ MIDDLEWARE = [
 CORS_ORIGIN_WHITELIST = ("http://*", "https://*")
 CORS_ALLOW_CREDENTIALS = True
 
-ROOT_URLCONF = 'TestProject.urls'
+ROOT_URLCONF = 'config.urls'
 
 TEMPLATES = [
     {
@@ -87,7 +87,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'TestProject.wsgi.application'
+WSGI_APPLICATION = 'config.wsgi.application'
 
 
 # Database
