+++ dontworry/config/urls.py
@@ -1,5 +1,5 @@
 """
-URL configuration for TestProject project.
+URL configuration for config project.
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/4.2/topics/http/urls/
